api/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.10.1
bleach==6.1.0
cachetools==5.3.3
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cssbeautifier==1.15.1
decorator==5.1.1
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
drf-yasg==1.21.7
EditorConfig==0.12.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
filelock==3.13.1
flake8==6.1.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
identify==2.5.36
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipython==8.14.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
Mako==1.3.0
Markdown==3.5.2
MarkupSafe==2.1.3
martor==1.6.44
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.11.14
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.7.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.0
pydantic==2.5.1
pydantic_core==2.14.3
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.2
pytest==7.4.3
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.4
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.5
uvicorn==0.24.0.post1
virtualenv==20.24.6
wcwidth==0.2.6
webencodings==0.5.1
